Extract turn engine and edit regenerator
