+++ src/mindroom/turn_engine.py
@@ -1,0 +1,878 @@
+"""Turn sequencing extracted from the Matrix bot runtime shell."""
+
+from __future__ import annotations
+
+import time
+from dataclasses import dataclass
+from typing import TYPE_CHECKING, Any, Protocol, cast
+
+import nio
+
+from mindroom import interactive
+from mindroom.attachments import merge_attachment_ids, parse_attachment_ids_from_event_source
+from mindroom.authorization import (
+    get_effective_sender_id_for_reply_permissions,
+    is_authorized_sender,
+)
+from mindroom.coalescing import (
+    CoalescedBatch,
+    CoalescingGate,
+    CoalescingKey,
+    PendingEvent,
+    PreparedTextEvent,
+    build_batch_dispatch_event,
+)
+from mindroom.commands.parsing import command_parser
+from mindroom.constants import (
+    ATTACHMENT_IDS_KEY,
+    ORIGINAL_SENDER_KEY,
+    ROUTER_AGENT_NAME,
+    STREAM_STATUS_KEY,
+    STREAM_STATUS_PENDING,
+    STREAM_STATUS_STREAMING,
+    VOICE_RAW_AUDIO_FALLBACK_KEY,
+    RuntimePaths,
+)
+from mindroom.delivery_gateway import SendTextRequest
+from mindroom.handled_turns import HandledTurnLedger, HandledTurnState
+from mindroom.hooks.ingress import (
+    hook_ingress_policy,
+    is_automation_source_kind,
+    should_handle_interactive_text_response,
+)
+from mindroom.inbound_turn_normalizer import (
+    BatchMediaAttachmentRequest,
+    DispatchPayload,
+    DispatchPayloadWithAttachmentsRequest,
+    InboundTurnNormalizer,
+    TextNormalizationRequest,
+    VoiceNormalizationRequest,
+)
+from mindroom.matrix.event_info import EventInfo
+from mindroom.matrix.identity import extract_agent_name, is_agent_id
+from mindroom.matrix.message_content import is_v2_sidecar_text_preview
+from mindroom.post_response_effects import matrix_run_metadata_for_handled_turn, record_handled_turn
+from mindroom.timing import (
+    attach_dispatch_pipeline_timing,
+    create_dispatch_pipeline_timing,
+    get_dispatch_pipeline_timing,
+)
+from mindroom.timing import timing_scope as timing_scope_context
+
+if TYPE_CHECKING:
+    from collections.abc import Sequence
+
+    import structlog
+    from agno.media import Image
+
+    from mindroom.bot_runtime_view import BotRuntimeView
+    from mindroom.conversation_resolver import ConversationResolver, MessageContext
+    from mindroom.conversation_state_writer import ConversationStateWriter
+    from mindroom.delivery_gateway import DeliveryGateway
+    from mindroom.dispatch_planner import DispatchPlanner, ResponseAction
+    from mindroom.history.types import HistoryScope
+    from mindroom.hooks import MessageEnvelope
+    from mindroom.matrix.client import ResolvedVisibleMessage
+    from mindroom.matrix.conversation_access import MatrixConversationAccess
+    from mindroom.matrix.identity import MatrixID
+    from mindroom.message_target import MessageTarget
+    from mindroom.response_coordinator import ResponseCoordinator
+
+type _MediaDispatchEvent = (
+    nio.RoomMessageImage
+    | nio.RoomEncryptedImage
+    | nio.RoomMessageFile
+    | nio.RoomEncryptedFile
+    | nio.RoomMessageVideo
+    | nio.RoomEncryptedVideo
+)
+type _InboundMediaEvent = _MediaDispatchEvent | nio.RoomMessageAudio | nio.RoomEncryptedAudio
+type _TextDispatchEvent = nio.RoomMessageText | PreparedTextEvent
+type _DispatchEvent = _TextDispatchEvent | _MediaDispatchEvent
+
+
+class _EditRegenerator(Protocol):
+    """Minimal edit-regeneration surface needed by turn sequencing."""
+
+    async def handle_message_edit(
+        self,
+        room: nio.MatrixRoom,
+        event: nio.RoomMessageText,
+        event_info: EventInfo,
+        requester_user_id: str,
+    ) -> None:
+        """Regenerate the owned response for one edited user turn."""
+
+
+@dataclass(frozen=True)
+class _PrecheckedEvent[T]:
+    """A raw or prepared event that already passed ingress prechecks."""
+
+    event: T
+    requester_user_id: str
+
+
+type _PrecheckedTextDispatchEvent = _PrecheckedEvent[_TextDispatchEvent]
+type _PrecheckedMediaDispatchEvent = _PrecheckedEvent[_MediaDispatchEvent]
+
+
+@dataclass(frozen=True)
+class TurnEngineDeps:
+    """Collaborators needed for Phase 1 turn sequencing."""
+
+    runtime: BotRuntimeView
+    logger: structlog.stdlib.BoundLogger
+    runtime_paths: RuntimePaths
+    agent_name: str
+    matrix_id: MatrixID
+    handled_turn_ledger: HandledTurnLedger
+    conversation_access: MatrixConversationAccess
+    resolver: ConversationResolver
+    normalizer: InboundTurnNormalizer
+    dispatch_planner: DispatchPlanner
+    response_coordinator: ResponseCoordinator
+    delivery_gateway: DeliveryGateway
+    state_writer: ConversationStateWriter
+    coalescing_gate: CoalescingGate
+    edit_regenerator: _EditRegenerator
+
+
+@dataclass
+class TurnEngine:
+    """Own sequencing for one inbound text or media turn."""
+
+    deps: TurnEngineDeps
+
+    def _client(self) -> nio.AsyncClient:
+        client = self.deps.runtime.client
+        if client is None:
+            msg = "Matrix client is not ready for turn execution"
+            raise RuntimeError(msg)
+        return client
+
+    def _requester_user_id(
+        self,
+        *,
+        sender: str,
+        source: object,
+    ) -> str:
+        """Return the effective requester for reply-permission checks."""
+        source_dict = cast("dict[str, Any] | None", source if isinstance(source, dict) else None)
+        content = source_dict.get("content") if source_dict is not None else None
+        if (
+            sender == self.deps.matrix_id.full_id
+            and isinstance(content, dict)
+            and isinstance(content.get(ORIGINAL_SENDER_KEY), str)
+        ):
+            return content[ORIGINAL_SENDER_KEY]
+        return get_effective_sender_id_for_reply_permissions(
+            sender,
+            source_dict,
+            self.deps.runtime.config,
+            self.deps.runtime_paths,
+        )
+
+    def _is_trusted_internal_relay_event(self, event: _DispatchEvent) -> bool:
+        """Return whether one agent-authored relay should bypass user-turn coalescing."""
+        if not isinstance(event, nio.RoomMessageText | PreparedTextEvent):
+            return False
+        if extract_agent_name(event.sender, self.deps.runtime.config, self.deps.runtime_paths) is None:
+            return False
+        content = event.source.get("content") if isinstance(event.source, dict) else None
+        if not isinstance(content, dict):
+            return False
+        if content.get("com.mindroom.source_kind") == "scheduled":
+            return False
+        original_sender = content.get(ORIGINAL_SENDER_KEY)
+        return isinstance(original_sender, str) and bool(original_sender)
+
+    def _precheck_event(
+        self,
+        room: nio.MatrixRoom,
+        event: _DispatchEvent | _InboundMediaEvent,
+        *,
+        is_edit: bool = False,
+    ) -> str | None:
+        """Run shared early-exit checks for inbound text and media events."""
+        content = event.source.get("content") if isinstance(event.source, dict) else None
+        source_kind = content.get("com.mindroom.source_kind") if isinstance(content, dict) else None
+        requester_user_id = self._requester_user_id(
+            sender=event.sender,
+            source=event.source,
+        )
+
+        if requester_user_id == self.deps.matrix_id.full_id and source_kind != "hook_dispatch":
+            return None
+
+        if not is_edit and self.deps.handled_turn_ledger.has_responded(event.event_id):
+            return None
+
+        if not is_authorized_sender(
+            requester_user_id,
+            self.deps.runtime.config,
+            room.room_id,
+            self.deps.runtime_paths,
+            room_alias=room.canonical_alias,
+        ):
+            self._mark_source_events_responded(HandledTurnState.from_source_event_id(event.event_id))
+            return None
+
+        if not self.deps.dispatch_planner.can_reply_to_sender(requester_user_id):
+            self._mark_source_events_responded(HandledTurnState.from_source_event_id(event.event_id))
+            return None
+
+        return requester_user_id
+
+    def _precheck_dispatch_event[T: _DispatchEvent](
+        self,
+        room: nio.MatrixRoom,
+        event: T,
+        *,
+        is_edit: bool = False,
+    ) -> _PrecheckedEvent[T] | None:
+        """Return a typed prechecked event for turn dispatch."""
+        requester_user_id = self._precheck_event(room, event, is_edit=is_edit)
+        if requester_user_id is None:
+            return None
+        return _PrecheckedEvent(event=event, requester_user_id=requester_user_id)
+
+    def _mark_source_events_responded(self, handled_turn: HandledTurnState) -> None:
+        """Mark one or more source events as handled by the same terminal outcome."""
+        record_handled_turn(self.deps.handled_turn_ledger, handled_turn)
+
+    def _response_history_scope_for_action(
+        self,
+        response_action: ResponseAction,
+    ) -> HistoryScope | None:
+        """Return the persisted history scope used by one response action."""
+        if response_action.kind == "team":
+            assert response_action.form_team is not None
+            return self.deps.state_writer.team_history_scope(response_action.form_team.eligible_members)
+        if response_action.kind == "individual":
+            return self.deps.state_writer.history_scope()
+        return None
+
+    def _handled_turn_with_response_context(
+        self,
+        handled_turn: HandledTurnState,
+        *,
+        history_scope: HistoryScope | None,
+        conversation_target: MessageTarget | None,
+    ) -> HandledTurnState:
+        """Attach the persisted regeneration context for one response."""
+        return handled_turn.with_response_context(
+            response_owner=self.deps.agent_name,
+            history_scope=history_scope,
+            conversation_target=conversation_target,
+        )
+
+    def _has_newer_unresponded_in_thread(
+        self,
+        event: _TextDispatchEvent,
+        requester_user_id: str,
+        thread_history: Sequence[ResolvedVisibleMessage],
+    ) -> bool:
+        """Return True when a newer unresponded message from the same sender exists."""
+        if isinstance(event, PreparedTextEvent) and is_automation_source_kind(event.source_kind_override or ""):
+            return False
+        event_ts = event.server_timestamp
+        if event_ts is None or not thread_history:
+            return False
+        for message in thread_history:
+            if message.sender != requester_user_id:
+                continue
+            if message.timestamp is None or message.timestamp <= event_ts:
+                continue
+            if message.event_id == event.event_id:
+                continue
+            if self.deps.handled_turn_ledger.has_responded(message.event_id):
+                continue
+            if (
+                message.body
+                and isinstance(message.body, str)
+                and command_parser.parse(message.body.strip()) is not None
+            ):
+                continue
+            self.deps.logger.info(
+                "Skipping older message — newer unresponded message from same sender in thread",
+                skipped_event_id=event.event_id,
+                newer_event_id=message.event_id,
+            )
+            return True
+        return False
+
+    def _should_skip_deep_synthetic_full_dispatch(
+        self,
+        *,
+        event_id: str,
+        envelope: MessageEnvelope,
+    ) -> bool:
+        """Return True when a deep synthetic hook relay must stop before dispatch."""
+        resolved_policy = hook_ingress_policy(envelope)
+        if resolved_policy.allow_full_dispatch:
+            return False
+        self.deps.logger.debug(
+            "Ignoring deep synthetic hook relay before command/response dispatch",
+            event_id=event_id,
+            source_kind=envelope.source_kind,
+            hook_source=envelope.hook_source,
+            message_received_depth=envelope.message_received_depth,
+        )
+        return True
+
+    def _should_bypass_coalescing_for_active_thread_follow_up(
+        self,
+        *,
+        target: MessageTarget,
+        source_kind: str,
+        sender_id: str,
+    ) -> bool:
+        """Return whether one human thread follow-up should skip in-flight coalescing."""
+        if target.resolved_thread_id is None:
+            return False
+        if is_automation_source_kind(source_kind):
+            return False
+        if is_agent_id(sender_id, self.deps.runtime.config, self.deps.runtime_paths):
+            return False
+        return self.deps.response_coordinator.has_active_response_for_target(target)
+
+    async def _coalescing_key_for_event(
+        self,
+        room: nio.MatrixRoom,
+        event: _DispatchEvent,
+        requester_user_id: str,
+    ) -> CoalescingKey:
+        """Return the sender or thread scoped dispatch key for one event."""
+        return (
+            room.room_id,
+            await self.deps.resolver.coalescing_thread_id(room, event),
+            requester_user_id,
+        )
+
+    async def _enqueue_for_dispatch(
+        self,
+        event: _DispatchEvent,
+        room: nio.MatrixRoom,
+        *,
+        source_kind: str,
+        requester_user_id: str | None = None,
+        coalescing_key: CoalescingKey | None = None,
+    ) -> None:
+        """Route one inbound event through the live coalescing gate."""
+        dispatch_timing = get_dispatch_pipeline_timing(event.source)
+        if dispatch_timing is not None:
+            dispatch_timing.mark("gate_enter")
+        effective_requester_user_id = requester_user_id or self._requester_user_id(
+            sender=event.sender,
+            source=event.source,
+        )
+        if self._is_trusted_internal_relay_event(event):
+            if dispatch_timing is not None:
+                dispatch_timing.note(coalescing_bypassed=True, coalescing_bypass_reason="trusted_internal_relay")
+                dispatch_timing.mark("gate_exit")
+            trusted_relay_event = cast("_TextDispatchEvent", event)
+            await self._dispatch_text_message(
+                room,
+                trusted_relay_event,
+                effective_requester_user_id,
+            )
+            return
+        await self.deps.coalescing_gate.enqueue(
+            coalescing_key or await self._coalescing_key_for_event(room, event, effective_requester_user_id),
+            PendingEvent(
+                event=event,
+                room=room,
+                source_kind=source_kind,
+            ),
+        )
+
+    async def _maybe_send_visible_voice_echo(
+        self,
+        room: nio.MatrixRoom,
+        event: nio.RoomMessageAudio | nio.RoomEncryptedAudio,
+        *,
+        text: str,
+        thread_id: str | None,
+    ) -> str | None:
+        """Optionally post a display-only router echo for normalized audio."""
+        if self.deps.agent_name != ROUTER_AGENT_NAME or not self.deps.runtime.config.voice.visible_router_echo:
+            return None
+
+        existing_visible_echo_event_id = self.deps.handled_turn_ledger.get_visible_echo_event_id(event.event_id)
+        if existing_visible_echo_event_id is not None:
+            return existing_visible_echo_event_id
+
+        target = self.deps.resolver.build_message_target(
+            room_id=room.room_id,
+            thread_id=thread_id,
+            reply_to_event_id=event.event_id,
+            event_source=event.source,
+        )
+        visible_echo_event_id = await self.deps.delivery_gateway.send_text(
+            SendTextRequest(
+                target=target,
+                response_text=text,
+                skip_mentions=True,
+            ),
+        )
+        if visible_echo_event_id is not None:
+            self.deps.handled_turn_ledger.record_visible_echo(event.event_id, visible_echo_event_id)
+        return visible_echo_event_id
+
+    async def handle_coalesced_batch(self, batch: CoalescedBatch) -> None:
+        """Dispatch one flushed batch through the normal text pipeline."""
+        dispatch_event = build_batch_dispatch_event(batch)
+        dispatch_timing = get_dispatch_pipeline_timing(dispatch_event.source)
+        if dispatch_timing is not None:
+            dispatch_timing.mark("gate_exit")
+        batch_coalescing_key = await self._coalescing_key_for_event(
+            batch.room,
+            batch.primary_event,
+            batch.requester_user_id,
+        )
+        canonical_key = (
+            batch.room.room_id,
+            self.deps.resolver.build_message_target(
+                room_id=batch.room.room_id,
+                thread_id=batch_coalescing_key[1],
+                reply_to_event_id=dispatch_event.event_id,
+                event_source=dispatch_event.source,
+            ).resolved_thread_id,
+            batch.requester_user_id,
+        )
+        self.deps.coalescing_gate.retarget(batch_coalescing_key, canonical_key)
+        async with self.deps.resolver.turn_thread_cache_scope():
+            await self._dispatch_text_message(
+                batch.room,
+                dispatch_event,
+                batch.requester_user_id,
+                media_events=batch.media_events or None,
+                handled_turn=HandledTurnState.create(
+                    batch.source_event_ids,
+                    source_event_prompts=batch.source_event_prompts,
+                ),
+            )
+
+    async def handle_text_event(self, room: nio.MatrixRoom, event: nio.RoomMessageText) -> None:
+        """Handle one inbound text event."""
+        async with self.deps.resolver.turn_thread_cache_scope():
+            await self._handle_message_inner(room, event)
+
+    async def _handle_message_inner(self, room: nio.MatrixRoom, event: nio.RoomMessageText) -> None:
+        """Handle one text message inside the per-turn thread-history cache scope."""
+        ingress_thread_id = await self.deps.resolver.coalescing_thread_id(room, event)
+        self.deps.logger.info(
+            "Received message",
+            event_id=event.event_id,
+            room_id=room.room_id,
+            sender=event.sender,
+            thread_id=ingress_thread_id,
+        )
+        dispatch_timing = create_dispatch_pipeline_timing(
+            event_id=event.event_id,
+            room_id=room.room_id,
+        )
+        attach_dispatch_pipeline_timing(event.source, dispatch_timing)
+        event_info = EventInfo.from_event(event.source)
+        await self.deps.conversation_access.append_live_event(room.room_id, event, event_info=event_info)
+        if not isinstance(event.body, str):
+            return
+        event_content = event.source.get("content") if isinstance(event.source, dict) else None
+        if isinstance(event_content, dict) and event_content.get(STREAM_STATUS_KEY) in {
+            STREAM_STATUS_PENDING,
+            STREAM_STATUS_STREAMING,
+        }:
+            return
+
+        prechecked_event = self._precheck_dispatch_event(room, event, is_edit=event_info.is_edit)
+        if prechecked_event is None:
+            return
+
+        if event_info.is_edit:
+            await self.deps.edit_regenerator.handle_message_edit(
+                room,
+                prechecked_event.event,
+                event_info,
+                prechecked_event.requester_user_id,
+            )
+            return
+
+        prepared_event = await self.deps.normalizer.resolve_text_event(
+            TextNormalizationRequest(event=prechecked_event.event),
+        )
+        attach_dispatch_pipeline_timing(prepared_event.source, dispatch_timing)
+        envelope = self.deps.resolver.build_ingress_envelope(
+            room_id=room.room_id,
+            event=prepared_event,
+            requester_user_id=prechecked_event.requester_user_id,
+        )
+        if self._should_skip_deep_synthetic_full_dispatch(
+            event_id=prepared_event.event_id,
+            envelope=envelope,
+        ):
+            return
+        if should_handle_interactive_text_response(envelope):
+            await interactive.handle_text_response(self._client(), room, prepared_event, self.deps.agent_name)
+        coalescing_thread_id = await self.deps.resolver.coalescing_thread_id(room, prepared_event)
+        target = self.deps.resolver.build_message_target(
+            room_id=room.room_id,
+            thread_id=coalescing_thread_id,
+            reply_to_event_id=prepared_event.event_id,
+            event_source=prepared_event.source,
+        )
+        if self._should_bypass_coalescing_for_active_thread_follow_up(
+            target=target,
+            source_kind=envelope.source_kind,
+            sender_id=prepared_event.sender,
+        ):
+            if dispatch_timing is not None:
+                dispatch_timing.mark("gate_enter")
+                dispatch_timing.note(
+                    coalescing_bypassed=True,
+                    coalescing_bypass_reason="active_thread_follow_up",
+                )
+                dispatch_timing.mark("gate_exit")
+            await self._dispatch_text_message(
+                room,
+                prepared_event,
+                prechecked_event.requester_user_id,
+            )
+            return
+        await self._enqueue_for_dispatch(
+            prechecked_event.event,
+            room,
+            source_kind="message",
+            requester_user_id=prechecked_event.requester_user_id,
+            coalescing_key=(room.room_id, coalescing_thread_id, prechecked_event.requester_user_id),
+        )
+
+    async def _dispatch_text_message(  # noqa: C901, PLR0912, PLR0915
+        self,
+        room: nio.MatrixRoom,
+        event: _TextDispatchEvent | _PrecheckedTextDispatchEvent,
+        requester_user_id: str | None = None,
+        *,
+        media_events: list[_MediaDispatchEvent] | None = None,
+        handled_turn: HandledTurnState | None = None,
+    ) -> None:
+        """Run the normal text or command dispatch pipeline for a prepared text event."""
+        raw_event: _TextDispatchEvent
+        if isinstance(event, _PrecheckedEvent):
+            requester_user_id = event.requester_user_id
+            raw_event = cast("_TextDispatchEvent", event.event)
+        else:
+            raw_event = event
+        if requester_user_id is None:
+            msg = "requester_user_id is required when dispatching a raw event"
+            raise TypeError(msg)
+        router_event: _DispatchEvent = raw_event
+        event = await self.deps.normalizer.resolve_text_event(
+            TextNormalizationRequest(event=raw_event),
+        )
+        dispatch_timing = get_dispatch_pipeline_timing(raw_event.source)
+        attach_dispatch_pipeline_timing(event.source, dispatch_timing)
+        timing_scope_token = timing_scope_context.set(event.event_id[:20] if event.event_id else "unknown")
+        try:
+            if dispatch_timing is not None:
+                dispatch_timing.mark("dispatch_start")
+            dispatch_started_at = time.monotonic()
+            handled_turn = handled_turn or HandledTurnState.from_source_event_id(event.event_id)
+
+            if dispatch_timing is not None:
+                dispatch_timing.mark("dispatch_prepare_start")
+            dispatch = await self.deps.dispatch_planner.prepare_dispatch(
+                room,
+                event,
+                requester_user_id,
+                event_label="message",
+                handled_turn=handled_turn,
+            )
+            if dispatch_timing is not None:
+                dispatch_timing.mark("dispatch_prepare_ready")
+            if dispatch is None:
+                return
+
+            command = command_parser.parse(event.body) if not media_events else None
+            if command:
+                if self.deps.agent_name == ROUTER_AGENT_NAME:
+                    await self.deps.dispatch_planner.execute_command(
+                        room=room,
+                        event=event,
+                        requester_user_id=requester_user_id,
+                        command=command,
+                        source_envelope=dispatch.envelope,
+                    )
+                return
+
+            if self._has_newer_unresponded_in_thread(
+                event,
+                requester_user_id,
+                dispatch.context.thread_history,
+            ):
+                self._mark_source_events_responded(handled_turn)
+                return
+            if self._should_skip_deep_synthetic_full_dispatch(
+                event_id=event.event_id,
+                envelope=dispatch.envelope,
+            ):
+                return
+            if dispatch.context.requires_full_thread_history:
+                await self.deps.resolver.hydrate_dispatch_context(room, event, dispatch.context)
+            content = event.source.get("content") if isinstance(event.source, dict) else None
+            message_attachment_ids = parse_attachment_ids_from_event_source(event.source)
+            message_extra_content: dict[str, Any] = {}
+            if message_attachment_ids:
+                message_extra_content[ATTACHMENT_IDS_KEY] = message_attachment_ids
+            if isinstance(content, dict):
+                original_sender = content.get(ORIGINAL_SENDER_KEY)
+                if isinstance(original_sender, str):
+                    message_extra_content[ORIGINAL_SENDER_KEY] = original_sender
+                raw_audio_fallback = content.get(VOICE_RAW_AUDIO_FALLBACK_KEY)
+                if isinstance(raw_audio_fallback, bool) and raw_audio_fallback:
+                    message_extra_content[VOICE_RAW_AUDIO_FALLBACK_KEY] = True
+            router_extra_content = dict(message_extra_content)
+            if media_events and ORIGINAL_SENDER_KEY not in router_extra_content:
+                router_extra_content[ORIGINAL_SENDER_KEY] = requester_user_id
+            if dispatch_timing is not None:
+                dispatch_timing.mark("dispatch_plan_start")
+            plan = await self.deps.dispatch_planner.plan_dispatch(
+                room,
+                event,
+                dispatch,
+                extra_content=router_extra_content or None,
+                media_events=media_events,
+                handled_turn=handled_turn,
+                router_event=media_events[0]
+                if media_events and len(handled_turn.source_event_ids) == 1
+                else router_event,
+            )
+            if dispatch_timing is not None:
+                dispatch_timing.mark("dispatch_plan_ready")
+            if plan.kind == "ignore":
+                if plan.handled_turn_outcome is not None:
+                    self._mark_source_events_responded(plan.handled_turn_outcome)
+                return
+            if plan.kind == "route":
+                route_event = plan.router_event or event
+                single_direct_media_route = (
+                    isinstance(
+                        route_event,
+                        nio.RoomMessageFile
+                        | nio.RoomEncryptedFile
+                        | nio.RoomMessageVideo
+                        | nio.RoomEncryptedVideo
+                        | nio.RoomMessageImage
+                        | nio.RoomEncryptedImage,
+                    )
+                    and media_events == [route_event]
+                    and handled_turn.source_event_ids == (event.event_id,)
+                )
+                routing_kwargs: dict[str, Any] = {
+                    "message": event.body if media_events else plan.router_message,
+                    "requester_user_id": dispatch.requester_user_id,
+                    "extra_content": plan.extra_content,
+                }
+                if plan.media_events is not None and not single_direct_media_route:
+                    routing_kwargs["media_events"] = plan.media_events
+                if (
+                    plan.handled_turn is not None
+                    and list(plan.handled_turn.source_event_ids) != [route_event.event_id]
+                    and not single_direct_media_route
+                ):
+                    routing_kwargs["handled_turn"] = self._handled_turn_with_response_context(
+                        plan.handled_turn,
+                        history_scope=None,
+                        conversation_target=dispatch.target,
+                    )
+                await self.deps.dispatch_planner.execute_router_relay(
+                    room,
+                    route_event,
+                    dispatch.context.thread_history,
+                    dispatch.context.thread_id,
+                    **routing_kwargs,
+                )
+                return
+            assert plan.response_action is not None
+            handled_turn = self._handled_turn_with_response_context(
+                handled_turn,
+                history_scope=self._response_history_scope_for_action(plan.response_action),
+                conversation_target=dispatch.target,
+            )
+            matrix_run_metadata = matrix_run_metadata_for_handled_turn(handled_turn)
+
+            async def build_payload(context: MessageContext) -> DispatchPayload:
+                effective_thread_id = self.deps.resolver.build_message_target(
+                    room_id=room.room_id,
+                    thread_id=context.thread_id,
+                    reply_to_event_id=event.event_id,
+                    event_source=event.source,
+                ).resolved_thread_id
+                media_attachment_ids: list[str] = []
+                fallback_images: list[Image] | None = None
+                if media_events:
+                    media_result = await self.deps.normalizer.register_batch_media_attachments(
+                        BatchMediaAttachmentRequest(
+                            room_id=room.room_id,
+                            thread_id=effective_thread_id,
+                            media_events=media_events,
+                        ),
+                    )
+                    media_attachment_ids = media_result.attachment_ids
+                    fallback_images = media_result.fallback_images
+                return await self.deps.normalizer.build_dispatch_payload_with_attachments(
+                    DispatchPayloadWithAttachmentsRequest(
+                        room_id=room.room_id,
+                        prompt=event.body,
+                        current_attachment_ids=merge_attachment_ids(
+                            message_attachment_ids,
+                            media_attachment_ids,
+                        ),
+                        thread_id=context.thread_id,
+                        media_thread_id=effective_thread_id,
+                        thread_history=context.thread_history,
+                        fallback_images=fallback_images,
+                    ),
+                )
+
+            await self.deps.dispatch_planner.execute_response_action(
+                room,
+                event,
+                dispatch,
+                plan.response_action,
+                build_payload,
+                processing_log="Processing",
+                dispatch_started_at=dispatch_started_at,
+                handled_turn=handled_turn,
+                matrix_run_metadata=matrix_run_metadata,
+            )
+        finally:
+            timing_scope_context.reset(timing_scope_token)
+
+    async def handle_media_event(
+        self,
+        room: nio.MatrixRoom,
+        event: _MediaDispatchEvent,
+    ) -> None:
+        """Handle one inbound media event."""
+        async with self.deps.resolver.turn_thread_cache_scope():
+            await self._handle_media_message_inner(room, event)
+
+    async def _handle_media_message_inner(
+        self,
+        room: nio.MatrixRoom,
+        event: _MediaDispatchEvent,
+    ) -> None:
+        """Handle one media event inside the per-turn thread-history cache scope."""
+        prechecked_event = self._precheck_dispatch_event(room, event)
+        if prechecked_event is None:
+            return
+
+        if await self._dispatch_special_media_as_text(room, prechecked_event):
+            return
+        event = prechecked_event.event
+        await self._enqueue_for_dispatch(
+            event,
+            room,
+            source_kind="image" if isinstance(event, nio.RoomMessageImage | nio.RoomEncryptedImage) else "media",
+            requester_user_id=prechecked_event.requester_user_id,
+        )
+
+    async def _dispatch_special_media_as_text(
+        self,
+        room: nio.MatrixRoom,
+        prechecked_event: _PrecheckedMediaDispatchEvent,
+    ) -> bool:
+        """Handle media events that normalize into the text dispatch pipeline."""
+        event = prechecked_event.event
+        if isinstance(event, nio.RoomMessageAudio | nio.RoomEncryptedAudio):
+            await self._on_audio_media_message(
+                room,
+                _PrecheckedEvent(
+                    event=event,
+                    requester_user_id=prechecked_event.requester_user_id,
+                ),
+            )
+            return True
+        if isinstance(event, nio.RoomMessageFile | nio.RoomEncryptedFile):
+            return await self._dispatch_file_sidecar_text_preview(
+                room,
+                _PrecheckedEvent(
+                    event=event,
+                    requester_user_id=prechecked_event.requester_user_id,
+                ),
+            )
+        return False
+
+    async def _on_audio_media_message(
+        self,
+        room: nio.MatrixRoom,
+        prechecked_event: _PrecheckedEvent[nio.RoomMessageAudio | nio.RoomEncryptedAudio],
+    ) -> None:
+        """Normalize audio into a synthetic text event and reuse text dispatch."""
+        event = prechecked_event.event
+
+        if is_agent_id(event.sender, self.deps.runtime.config, self.deps.runtime_paths):
+            self.deps.logger.debug(
+                "Ignoring agent audio event for voice transcription",
+                event_id=event.event_id,
+                sender=event.sender,
+            )
+            self._mark_source_events_responded(HandledTurnState.from_source_event_id(event.event_id))
+            return
+
+        normalized_voice = await self.deps.normalizer.prepare_voice_event(
+            VoiceNormalizationRequest(
+                room=room,
+                event=event,
+            ),
+        )
+        if normalized_voice is None:
+            self._mark_source_events_responded(HandledTurnState.from_source_event_id(event.event_id))
+            return
+
+        await self._maybe_send_visible_voice_echo(
+            room,
+            event,
+            text=normalized_voice.event.body,
+            thread_id=normalized_voice.effective_thread_id,
+        )
+
+        await self._enqueue_for_dispatch(
+            normalized_voice.event,
+            room,
+            source_kind="voice",
+            requester_user_id=prechecked_event.requester_user_id,
+        )
+
+    async def _dispatch_file_sidecar_text_preview(
+        self,
+        room: nio.MatrixRoom,
+        prechecked_event: _PrecheckedEvent[nio.RoomMessageFile | nio.RoomEncryptedFile],
+    ) -> bool:
+        """Dispatch one sidecar-backed file preview through the normal text pipeline."""
+        event = prechecked_event.event
+        if not is_v2_sidecar_text_preview(event.source):
+            return False
+
+        prepared_text_event = await self.deps.normalizer.prepare_file_sidecar_text_event(event)
+        assert prepared_text_event is not None
+        envelope = self.deps.resolver.build_ingress_envelope(
+            room_id=room.room_id,
+            event=prepared_text_event,
+            requester_user_id=prechecked_event.requester_user_id,
+        )
+        if self._should_skip_deep_synthetic_full_dispatch(
+            event_id=prepared_text_event.event_id,
+            envelope=envelope,
+        ):
+            return True
+        if should_handle_interactive_text_response(envelope):
+            await interactive.handle_text_response(self._client(), room, prepared_text_event, self.deps.agent_name)
+        await self._dispatch_text_message(
+            room,
+            _PrecheckedEvent(
+                event=prepared_text_event,
+                requester_user_id=prechecked_event.requester_user_id,
+            ),
+        )
+        return True
