+++ tests/test_edit_after_restart.py
@@ -11,8 +11,9 @@ import pytest
 from mindroom.bot import AgentBot
 from mindroom.constants import resolve_runtime_paths
 from mindroom.handled_turns import HandledTurnLedger, HandledTurnState
+from mindroom.matrix.identity import MatrixID
 from mindroom.matrix.users import AgentMatrixUser
-from tests.conftest import wrap_extracted_collaborators
+from tests.conftest import replace_turn_engine_deps, wrap_extracted_collaborators
 
 if TYPE_CHECKING:
     from pathlib import Path
@@ -58,7 +59,7 @@ async def test_bot_handles_redelivered_edit_after_restart(tmp_path: Path) -> Non
     config = Mock()
     config.agents = {"test_agent": Mock()}
     config.get_agent_knowledge_base_ids.return_value = []
-    config.get_ids.return_value = {}
+    config.get_ids.return_value = {"test_agent": MatrixID.parse("@test_agent:example.com")}
     config.get_mindroom_user_id.return_value = "@mindroom:example.com"
     config.authorization.agent_reply_permissions = {}
 
@@ -85,6 +86,8 @@ async def test_bot_handles_redelivered_edit_after_restart(tmp_path: Path) -> Non
 
     # Mock logger
     bot.logger = MagicMock()
+    replace_turn_engine_deps(bot, handled_turn_ledger=bot.handled_turn_ledger, logger=bot.logger)
+    replace_turn_engine_deps(bot, handled_turn_ledger=bot.handled_turn_ledger, logger=bot.logger)
 
     # Create a room
     room = nio.MatrixRoom(room_id="!test:example.com", own_user_id="@test_agent:example.com")
@@ -140,8 +143,8 @@ async def test_bot_handles_redelivered_edit_after_restart(tmp_path: Path) -> Non
 
     # Mock the methods needed for regeneration
     with (
-        patch.object(bot, "_handle_message_edit", new_callable=AsyncMock) as mock_handle_edit,
-        patch("mindroom.bot.is_authorized_sender", return_value=True),
+        patch.object(bot._edit_regenerator, "handle_message_edit", new_callable=AsyncMock) as mock_handle_edit,
+        patch("mindroom.turn_engine.is_authorized_sender", return_value=True),
     ):
         # Process the redelivered edit event
         await bot._on_message(room, edit_event)
@@ -169,7 +172,7 @@ async def test_bot_skips_duplicate_regular_message_after_restart(tmp_path: Path)
     config = Mock()
     config.agents = {"test_agent": Mock()}
     config.get_agent_knowledge_base_ids.return_value = []
-    config.get_ids.return_value = {}
+    config.get_ids.return_value = {"test_agent": MatrixID.parse("@test_agent:example.com")}
     config.get_mindroom_user_id.return_value = "@mindroom:example.com"
     config.authorization.agent_reply_permissions = {}
 
@@ -196,6 +199,7 @@ async def test_bot_skips_duplicate_regular_message_after_restart(tmp_path: Path)
 
     # Mock logger
     bot.logger = MagicMock()
+    replace_turn_engine_deps(bot, handled_turn_ledger=bot.handled_turn_ledger, logger=bot.logger)
 
     # Create a room
     room = nio.MatrixRoom(room_id="!test:example.com", own_user_id="@test_agent:example.com")
@@ -229,8 +233,8 @@ async def test_bot_skips_duplicate_regular_message_after_restart(tmp_path: Path)
 
     # Mock methods
     with (
-        patch.object(bot, "_dispatch_text_message", new_callable=AsyncMock) as mock_dispatch,
-        patch("mindroom.bot.is_authorized_sender", return_value=True),
+        patch.object(bot._turn_engine, "_dispatch_text_message", new_callable=AsyncMock) as mock_dispatch,
+        patch("mindroom.turn_engine.is_authorized_sender", return_value=True),
     ):
         # Process the redelivered message
         await bot._on_message(room, message_event)
