+++ tests/test_voice_bot_threading.py
@@ -20,6 +20,7 @@ from tests.conftest import (
     bind_runtime_paths,
     install_generate_response_mock,
     replace_dispatch_planner_deps,
+    replace_turn_engine_deps,
     runtime_paths_for,
     sync_bot_runtime_state,
     test_runtime_paths,
@@ -42,6 +43,7 @@ def _agent_bot(*, agent_user: AgentMatrixUser, storage_path: Path, config: Confi
 @pytest.fixture
 def mock_home_bot() -> AgentBot:
     """Create a single-agent bot for audio threading tests."""
+    tmpdir = Path(tempfile.mkdtemp())
     agent_user = AgentMatrixUser(
         agent_name="home",
         user_id="@mindroom_home:localhost",
@@ -53,9 +55,8 @@ def mock_home_bot() -> AgentBot:
         agents={"home": {"display_name": "HomeAssistant", "rooms": ["!test:server"]}},
         authorization={"default_room_access": True},
     )
-    config = bind_runtime_paths(config, test_runtime_paths(Path(tempfile.mkdtemp())))
-    with tempfile.TemporaryDirectory() as tmpdir:
-        bot = _agent_bot(agent_user=agent_user, storage_path=Path(tmpdir), config=config, rooms=["!test:server"])
+    config = bind_runtime_paths(config, test_runtime_paths(tmpdir))
+    bot = _agent_bot(agent_user=agent_user, storage_path=tmpdir, config=config, rooms=["!test:server"])
     wrap_extracted_collaborators(bot)
     bot.client = AsyncMock()
     bot.client.rooms = {}
@@ -64,6 +65,7 @@ def mock_home_bot() -> AgentBot:
     bot.logger = MagicMock()
     bot.handled_turn_ledger = MagicMock()
     bot.handled_turn_ledger.has_responded.return_value = False
+    replace_turn_engine_deps(bot, handled_turn_ledger=bot.handled_turn_ledger, logger=bot.logger)
     replace_dispatch_planner_deps(bot, handled_turn_ledger=bot.handled_turn_ledger)
     bot._generate_response = AsyncMock(return_value="$response")
     install_generate_response_mock(bot, bot._generate_response)
@@ -106,7 +108,7 @@ async def test_voice_message_in_main_room_creates_thread(mock_home_bot: AgentBot
     with (
         patch("mindroom.voice_handler._download_audio", new_callable=AsyncMock) as mock_download_audio,
         patch("mindroom.voice_handler._handle_voice_message", return_value="🎤 what is the weather"),
-        patch("mindroom.bot.is_authorized_sender", return_value=True),
+        patch("mindroom.turn_engine.is_authorized_sender", return_value=True),
         patch("mindroom.dispatch_planner.is_dm_room", new_callable=AsyncMock, return_value=False),
     ):
         mock_download_audio.return_value = Audio(content=b"voice-bytes", mime_type="audio/ogg")
@@ -147,7 +149,7 @@ async def test_voice_message_in_thread_continues_thread(mock_home_bot: AgentBot)
     with (
         patch("mindroom.voice_handler._download_audio", new_callable=AsyncMock) as mock_download_audio,
         patch("mindroom.voice_handler._handle_voice_message", return_value="🎤 show me the forecast"),
-        patch("mindroom.bot.is_authorized_sender", return_value=True),
+        patch("mindroom.turn_engine.is_authorized_sender", return_value=True),
         patch("mindroom.dispatch_planner.is_dm_room", new_callable=AsyncMock, return_value=False),
     ):
         mock_download_audio.return_value = Audio(content=b"voice-bytes", mime_type="audio/ogg")
@@ -203,7 +205,7 @@ async def test_voice_plain_reply_to_thread_message_uses_thread_root(mock_home_bo
     with (
         patch("mindroom.voice_handler._download_audio", new_callable=AsyncMock) as mock_download_audio,
         patch("mindroom.voice_handler._handle_voice_message", return_value="🎤 continue the same thread"),
-        patch("mindroom.bot.is_authorized_sender", return_value=True),
+        patch("mindroom.turn_engine.is_authorized_sender", return_value=True),
         patch("mindroom.dispatch_planner.is_dm_room", new_callable=AsyncMock, return_value=False),
     ):
         mock_download_audio.return_value = Audio(content=b"voice-bytes", mime_type="audio/ogg")
